CHORE: Move graffix requests content board to /backoffice/graffix-requests/[id]. Set original page to a message to user to navigate to new page
