+++ src/pages/backoffice/index.tsx
@@ -1,0 +1,14 @@
+import { Page } from 'modules';
+import Head from 'next/head';
+
+export default function Backoffice() {
+  return (
+    <Page>
+      <Head>
+        <title>U-SU Backoffice</title>
+        <meta name="robots" content="noindex,nofollow" />
+      </Head>
+      Hello Backoffice!
+    </Page>
+  );
+}
